Moved unit tests to right class
